requirements.txt
streamlit
pandas
numpy
plotly
sentence-transformers
faiss-cpu
llama-cpp-python
scikit-learn
